src/requirements.txt
altgraph==0.17.4
asgiref==3.7.2
bleach==6.0.0
bottle==0.12.25
bottle-websocket==0.2.9
configparser==6.0.0
Django==4.2.4
django-sass==1.1.0
Eel==0.16.0
ffmpeg-python==0.2.0
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.0
libsass==0.22.0
Markdown==3.4.4
numpy==1.25.2
packaging==23.2
pandas==2.1.0
Pillow==10.0.0
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
pyjsparser==2.7.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tinycss2==1.1.1
tzdata==2023.3
tzlocal==5.0.1
webencodings==0.5.1
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.0
